% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{checkVarTab}
\alias{checkVarTab}
\title{Check the variable in a dataset.}
\usage{
checkVarTab(data, subset, grouping)
}
\arguments{
\item{data}{a data frame}

\item{subset}{a subset variable should be a binomial fctor}

\item{grouping}{a binomial factor variable in the data frame to divide into two groups}
}
\description{
Check the variable in a dataset.
}
