% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{ChiF}
\alias{ChiF}
\title{This function takes grouping vector and a DATA FRame ONLY OF  variables to test}
\usage{
ChiF(variablesDF, grouping)
}
\description{
This function takes grouping vector and a DATA FRame ONLY OF  variables to test
}
