% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{funChi}
\alias{funChi}
\title{This function makes a chisq test on selected variables given a grouping variable.
It outputs a p.value}
\usage{
funChi(var, grouping_var)
}
\description{
This function makes a chisq test on selected variables given a grouping variable.
It outputs a p.value
}
