% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{makeTests}
\alias{makeTests}
\title{Do a bunch of statistial tests on a data frame
This function automatically chooses test and performs them on the anaphylaxis registry database}
\usage{
makeTests(grouping, rdb)
}
\arguments{
\item{grouping}{Factor with two levels to group the cases into case control sets}

\item{rdb}{database on which to perform the test (you can restrict it)}
}
\value{
Tbl
}
\description{
Do a bunch of statistial tests on a data frame
This function automatically chooses test and performs them on the anaphylaxis registry database
}
